CompositeNetworkEditPart wrongly uses EContentAdapter
The CompositeNetworkEditPart was attaching an EContentAdapter to all its
contained elements. As this class is also used for typed Subapps this
could lead to major performance issues.

With this commit the code that is used for untyped subapps is also used
for CompositeNetworkEditParts.
